Fix tooltip and screen resize of bar graph
- filled in the percentage data in the hover tooltip when the bar graph
for a particular nutrient is switched to percentage

- allowed the SVG in both bar graph and sunburst graph to be
responsive for any screen width

- organized files to follow MVC design pattern
