+++ assets/colours/graphColours.js
@@ -1,0 +1,14 @@
+export const graphColours = {
+    "Baby Foods": "#BAABDA",
+    "Beverages (Excluding Milks)": "#F8CB2E",
+    "Dairy & Plant-Based Beverages": "#39B7E3",
+    "Fats & Oils": "#D8B384",
+    "Fish & Seafood": "#FF6969",
+    "Fruits & Vegetables": "#3CCF4E",
+    "Grain Products": "#EF5B0C",
+    "Meat & Poultry": "#BB2525",
+    "Meat Alternatives": "#482121",
+    "Nutritional Beverages & Bars": "#84F1CD",
+    "Soups - Sauces - Spices & Other Ingredients": "#005DD0",
+    "Sweets - Sugars & Savoury Snacks": "#824BD0"
+}
