+++ assets/dimensions/graphDimensions.js
@@ -1,0 +1,30 @@
+export const dims = Object.freeze({
+    upperGraphWidth: 800,
+    upperGraphHeight: 600,
+    upperGraphLeft: 60,
+    upperGraphRight: 400,
+    upperGraphTop: 60,
+    upperGraphBottom: 60,
+    upperGraphTooltipMinWidth: 140,
+    upperGraphInfoBoxWidth: 240,
+    upperGraphInfoBoxHeight: 200,
+    upperGraphInfoBoxFontSize: 12,
+    upperGraphChartHeadingFontSize: 20,
+    lowerGraphWidth: 700,
+    lowerGraphHeight: 700,
+    lowerGraphLeft: 60,
+    lowerGraphRight: 400,
+    lowerGraphTop: 60,
+    lowerGraphBottom: 60,
+    lowerGraphArcRadius: 70,
+    lowerGraphTooltipMinWidth: 140,
+    lowerGraphInfoBoxWidth: 240,
+    lowerGraphInfoBoxHeight: 200,
+    lowerGraphArcLabelFontSize: 12,
+    lowerGraphChartHeadingFontSize: 20,    
+    lowerGraphInfoBoxFontSize: 12,
+    legendFontSize: 12,
+    legendSquareSize: 12,
+    legendRowHeight: 20,
+    tableSectionBorderLeft: "1px solid black",
+})
