+++ model/dataModels/csvDataModel.js
@@ -1,0 +1,5 @@
+export class CSVDataModel {
+    constructor(data) {
+        this.data = data;
+    }
+}
