+++ styles/styles.css
@@ -1,0 +1,11 @@
+.graphExtra {
+    margin: 10px 0 0 0;
+}
+
+.lowerGraphExtra {
+    float: right;
+}
+
+#lowerGraphFilterGroupsButton {
+    position:relative;
+}
